Move chown function to action/mod.rs file
This will allow other actions to use this function
to change file permissons
